[S09P31A202-284] refactor: item 도메인 컬럼 자료형 변경 Integer -> Long

+++ backend/zupzup/src/main/java/com/twoez/zupzup/item/controller/dto/response/ItemResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 
 @Builder
 public record ItemResponse(
-        Long id, String name, String description, Integer exp, Integer price, String itemImgUrl) {
+        Long id, String name, String description, Integer exp, Long price, String itemImgUrl) {
     public static ItemResponse from(Item item) {
         return ItemResponse.builder()
                 .id(item.getId())
